Separate accessors and DB struct logic
